[COMMON] Mapper에 필요한 dependency 추가

See merge request s07-final/S07P31A604!5

+++ gardenary-be/src/main/java/com/gardenary/domain/exp/repository/ExpRepository.java
@@ -3,5 +3,5 @@ package com.gardenary.domain.exp.repository;
 import com.gardenary.domain.exp.entity.Exp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpRepository extends JpaRepository<Exp, Long> {
+public interface ExpRepository extends JpaRepository<Exp, Integer> {
 }
