+++ gardenary-be/src/main/java/com/gardenary/domain/flower/repository/QuestionAnswerRepository.java
@@ -3,5 +3,5 @@ package com.gardenary.domain.flower.repository;
 import com.gardenary.domain.flower.entity.QuestionAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionAnswerRepository extends JpaRepository<QuestionAnswer, Long> {
+public interface QuestionAnswerRepository extends JpaRepository<QuestionAnswer, Integer> {
 }
